server/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
distlib==0.3.7
docx2txt==0.8
filelock==3.12.4
Flask==2.3.3
Flask-Cors==4.0.0
flask-mysql-connector==1.1.0
frozenlist==1.4.0
google-auth==2.23.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
mysql-connector-python==8.1.0
numpy==1.26.0
oauthlib==3.2.2
openai==0.28.1
pandas==2.1.1
Pillow==10.1.0
platformdirs==3.10.0
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.23
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.5
virtualenv==20.24.5
Werkzeug==2.3.7
XlsxWriter==3.1.9
yarl==1.9.2